+++ src/mongoChangeStream.test.ts
@@ -5,6 +5,7 @@ import _ from 'lodash/fp'
 import { test } from 'node:test'
 import assert from 'node:assert'
 import { initSync } from './mongoChangeStream.js'
+import { JSONSchema, SyncOptions } from './types.js'
 import {
   ChangeStreamDocument,
   ChangeStreamInsertDocument,
@@ -16,15 +17,15 @@ import { faker } from '@faker-js/faker'
 import ms from 'ms'
 import { setTimeout } from 'node:timers/promises'
 
-const init = _.memoize(async () => {
+const init = _.memoize(async (options?: SyncOptions) => {
   const redis = new Redis({ keyPrefix: 'testing:' })
   const client = await MongoClient.connect(process.env.MONGO_CONN as string)
   const db = client.db()
   const coll = db.collection('testing')
-  const sync = initSync(redis, coll)
+  const sync = initSync(redis, coll, options)
   sync.emitter.on('stateChange', console.log)
 
-  return { sync, coll }
+  return { sync, db, coll, redis }
 })
 
 const genUser = () => ({
@@ -34,6 +35,20 @@ const genUser = () => ({
   zipCode: faker.address.zipCode(),
   createdAt: faker.date.past(),
 })
+
+const schema: JSONSchema = {
+  bsonType: 'object',
+  additionalProperties: false,
+  required: ['name'],
+  properties: {
+    _id: { bsonType: 'objectId' },
+    name: { bsonType: 'string' },
+    city: { bsonType: 'string' },
+    state: { bsonType: 'string' },
+    zipCode: { bsonType: 'string' },
+    createdAt: { bsonType: 'date' },
+  },
+}
 
 const numDocs = 500
 
@@ -95,6 +110,12 @@ test('initial scan should resume properly', async () => {
 
 test('should process records via change stream', async () => {
   const { sync, coll } = await init()
+  // Reset state
+  await sync.reset()
+  await coll.deleteMany({})
+
+  await populateCollection(coll)
+
   const processed = []
   const processRecord = async (doc: ChangeStreamDocument) => {
     await setTimeout(5)
@@ -115,6 +136,12 @@ test('should process records via change stream', async () => {
 
 test('change stream should resume properly', async () => {
   const { sync, coll } = await init()
+  // Reset state
+  await sync.reset()
+  await coll.deleteMany({})
+
+  await populateCollection(coll)
+
   const processed = []
   // Change stream
   const processRecord = async (doc: ChangeStreamDocument) => {
@@ -133,6 +160,7 @@ test('change stream should resume properly', async () => {
   await changeStream.stop()
   // Resume change stream
   changeStream.start()
+  // Wait for all documents to be processed
   await setTimeout(ms('10s'))
   // All change stream docs were processed
   assert.equal(processed.length, numDocs)
@@ -193,4 +221,87 @@ test('starting initial scan is idempotent', async () => {
   initialScan.start()
   initialScan.start()
   await initialScan.stop()
+})
+
+test('Should resync when resync flag is set', async () => {
+  const { sync, coll, redis } = await init()
+  // Reset state
+  await sync.reset()
+  await coll.deleteMany({})
+
+  let resyncTriggered = false
+  const processed = []
+
+  await populateCollection(coll)
+
+  const processRecords = async (docs: ChangeStreamInsertDocument[]) => {
+    await setTimeout(50)
+    processed.push(...docs)
+  }
+
+  const initialScan = await sync.runInitialScan(processRecords)
+  const resync = sync.detectResync()
+  initialScan.start()
+  resync.start()
+  // Allow for initial scan to start
+  await setTimeout(500)
+  // Trigger resync
+  await redis.set(sync.keys.resyncKey, 1)
+
+  sync.emitter.on('resync', async () => {
+    resyncTriggered = true
+    // Stop the initial scan
+    await initialScan.stop()
+    // Reset keys
+    await sync.reset()
+    // Reset processed
+    processed.length = 0
+    // Start initial scan
+    initialScan.start()
+  })
+  // Wait for initial scan to complete
+  await setTimeout(ms('5s'))
+  assert.ok(resyncTriggered)
+  assert.equal(processed.length, numDocs)
+  await initialScan.stop()
+  resync.stop()
+})
+
+test('Resync start/stop is idempotent', async () => {
+  const { sync } = await init()
+
+  const resync = sync.detectResync()
+  resync.start()
+  resync.start()
+  resync.stop()
+  resync.stop()
+})
+
+test('Detect schema change', async () => {
+  const { db, coll, sync } = await init()
+  // Set schema
+  await db.command({
+    collMod: coll.collectionName,
+    validator: { $jsonSchema: schema },
+  })
+  // Look for a new schema every 250 ms
+  const schemaChange = await sync.detectSchemaChange(db, {
+    shouldRemoveMetadata: true,
+    interval: 250,
+  })
+  let newSchema: object = {}
+  sync.emitter.on('schemaChange', ({ currentSchema }) => {
+    newSchema = currentSchema
+  })
+  // Start detecting schema changes
+  schemaChange.start()
+  // Modify the schema
+  schema.properties.email = { bsonType: 'string' }
+  await db.command({
+    collMod: coll.collectionName,
+    validator: { $jsonSchema: schema },
+  })
+  await setTimeout(ms('1s'))
+  assert.deepEqual(schema, newSchema)
+  schemaChange.stop()
 })